moving some of the instances to their right place
